Added apple data structure
Adds an apple data structure and a list of apples to the game. The
spawning of the apple should be random but really isn't (I should take a
look at this some other day). Still, I tried to add an `StdGen` pure
field in the Game, but apparently something is not working well with it.